% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{Access}
\alias{Access}
\title{Access member variable from all elements of a list}
\usage{
Access(x, name, unlist = TRUE)
}
\arguments{
\item{x}{A list.}

\item{name}{Name of variable to access from each element of the list.}

\item{unlist}{If \code{TRU}, unlist the result before returning.}
}
\description{
Access member variable from all elements of a list
}
\examples{
x = list()
x[[1]] = list(a = 1:10, b = 1:5)
x[[2]] = list(a = 1:5, b = 0, c = -1)
Access(x, "a")
Access(x, "a", unlist = FALSE)
Access(x, "b")
Access(x, "c")

}
