% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\alias{adapt_midpoint}
\title{Adapt}
\usage{
adapt_midpoint(h, N, control = rejection_control(), method = "greedy")
}
\arguments{
\item{h}{An FMM proposal}

\item{N}{Number of additional mixture components after adaptation.}

\item{control}{A control object from \code{rejection_control}}

\item{method}{Either \code{greedy} or \code{pps}. The \code{greedy} method
always chooses to bifrucate the region with the largest contribution to the
probability of rejection. The \code{pps} method draws randomly, with
probability proportion to that contribution.}
}
\description{
Adapt an FMM proposal using a midpoint rule-of-thumb.
}
