% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt-midpoint.R
\name{adapt_midpoint}
\alias{adapt_midpoint}
\title{Adapt Using Midpoint}
\usage{
adapt_midpoint(h, N, method = "pps", control = rejection_control())
}
\arguments{
\item{h}{An FMM proposal}

\item{N}{Number of additional mixture components after adaptation.}

\item{method}{Either \code{greedy} or \code{pps}. The \code{greedy} method
always chooses to bifrucate the region with the largest contribution to the
probability of rejection. The \code{pps} method draws randomly, with
probability proportion to that contribution.}

\item{control}{A control object from \code{rejection_control}}
}
\description{
Adapt an FMM proposal using a midpoint rule-of-thumb.
}
\examples{
g = normal_univariate_helper(0, 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }
support = univariate_const_region(-Inf, 10, w, g)
regions = list(support)

# Adapt the proposal
h_init = fmm_proposal(regions)
adapt_out = adapt_midpoint(h_init, N = 100, control = rejection_control(report = 1))

# Create a finite mixture proposal
h = adapt_out$h
h$rejection_bound()
h$rejection_bound(byregion = TRUE)

out = rejection(h, n = 200)
print(out |> unlist())

out = rejection(h, n = 200, rejection_control(extra_outputs = TRUE))
print(out$draws |> unlist())
print(out$rejects)

}
