% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect.R
\name{bounded}
\alias{bounded}
\alias{rect}
\alias{inv_rect}
\title{Rectangular transformation}
\usage{
rect(z, a, b)

inv_rect(x, a, b)
}
\arguments{
\item{z}{A point in the rectangle \eqn{[a_1,b_1] \times \cdots \times [a_d,b_d]}.}

\item{a}{A vector \eqn{(a_1, \ldots, a_d)}, Elements may be \code{-Inf}.}

\item{b}{A vector \eqn{(b_1, \ldots, b_d)}, Elements may be \code{Inf}.}

\item{x}{A point in \eqn{\mathbb{R}^{d}}.}
}
\description{
A transformation from Euclidean to a rectangle in \eqn{\mathbb{R}^n}, and
its inverse transformation.
}
\examples{
x = seq(-1, 1, length.out = 3)
a = rep(0, 3)
b = rep(1, 3)
z = inv_rect(x, a, b)
rect(z, a, b)

a = c(-Inf, 0, -Inf)
b = c(Inf, 1, Inf)
z = inv_rect(x, a, b)
rect(z, a, b)
}
