% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categ.R
\name{categ}
\alias{categ}
\alias{r_categ}
\title{Categorical Distribution}
\usage{
r_categ(n, p, log_p = FALSE)
}
\arguments{
\item{n}{Number of desired draws}

\item{p}{A \eqn{k}-dimensional vector of probabilities}

\item{log_p}{If \code{TRUE}, interpret given probabilities as having been
supplied on the log-scale. Otherwise, interpret them on the original scale.}
}
\value{
a vector of category indices whose elements are in
\eqn{1, \ldots, k}.
}
\description{
Sample from a matrix of categorical probabilities.
}
\details{
We make use of the Gumbel trick to draw from probabilities given on the
log-scale without having to normalize. Note that \code{r_categ} can be slow
with large \code{n} because it runs in a loop in plain R.
}
