% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal-spaced-knots.R
\name{equal_spaced_knots}
\alias{equal_spaced_knots}
\title{Equally Spaced Univariate Knots}
\usage{
equal_spaced_knots(N, lo, hi)
}
\arguments{
\item{N}{TBD}

\item{lo}{TBD}

\item{hi}{TBD}
}
\value{
TBD
}
\description{
Prepare \eqn{N} equally-spaced intervals between finite endpoints \code{lo} and
\code{hi}.
}
\examples{
print("TBD")

}
