% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMMProposal.R
\name{fmm_proposal}
\alias{fmm_proposal}
\title{fmm_proposal}
\usage{
fmm_proposal(regions)
}
\arguments{
\item{regions}{A list of regions that form a partition of the support.}
}
\description{
A more friendly constructor for \code{FMMProposal}
}
\examples{
# Define base distribution and weight function
g = normal_univariate_helper(mean = 0, sd = 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }

# Set up support
support = univariate_const_region(-Inf, 10, w, g)
regions = support$bifurcate()

# Create a finite mixture with one component
fmm = fmm_proposal(regions)
print(fmm)

}
