% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GammaTrunc}
\alias{GammaTrunc}
\alias{r_gamma_trunc}
\alias{d_gamma_trunc}
\alias{p_gamma_trunc}
\alias{q_gamma_trunc}
\title{Truncated Gamma Distribution}
\usage{
r_gamma_trunc(n, shape, rate, a, b)

d_gamma_trunc(x, shape, rate, a, b, log = FALSE)

p_gamma_trunc(q, shape, rate, a, b, lower = TRUE, log = FALSE)

q_gamma_trunc(p, shape, rate, a, b, lower = TRUE, log = FALSE)
}
\arguments{
\item{n}{Number of draws.}

\item{shape}{Shape parameter.}

\item{rate}{Rate parameter.}

\item{a}{lower limit of support.}

\item{b}{upper limit of support.}

\item{x}{Vector; argument of density.}

\item{log}{If \code{TRUE}, return densities and probabilities on the log-scale.}

\item{q}{Vector; argument of cumulative distribution function.}

\item{lower}{logical; if \code{TRUE} (default), probabilities are
\eqn{P(X \leq x)}; otherwise, \eqn{P(X > x)}.}

\item{p}{Vector; argument of quantile function.}
}
\value{
\code{d_gamma_trunc} computes the density, \code{r_gamma_trunc} generates random
deviates, \code{p_gamma_trunc} computes the CDF, and \code{q_gamma_trunc} computes
quantiles.
}
\description{
Functions for the Truncated Gamma distribution parameterized by shape and
rate, whose density is
\deqn{
f(x \mid a, b) =
\frac{
b^a
}{
\Gamma(a)
}
x^{-a-1} e^{-b / x} \textrm{I}(x \geq 0)
}
}
\examples{
shape = 10
rate = 5
a = 2
b = 5
x = r_gamma_trunc(100000, shape, rate, a, b)
xx = seq(a, b, length.out = 100)

plot(density(x))
lines(xx, d_gamma_trunc(xx, shape, rate, a, b), lty = 2, col = "blue", lwd = 2)

plot(ecdf(x))
lines(xx, p_gamma_trunc(xx, shape, rate, a, b), lty = 2, col = "blue", lwd = 2)

pp = seq(0, 1, length.out = 102) |> head(-1) |> tail(-1)
qq = quantile(x, probs = pp)
plot(pp, qq)
lines(pp, q_gamma_trunc(pp, shape, rate, a, b), lty = 2, col = "blue", lwd = 2)

}
