% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxreg.R
\name{getxreg}
\alias{getxreg}
\title{get the xreg from a vws regression}
\usage{
getxreg(object, data = NULL)
}
\arguments{
\item{object}{an object of class \code{vws}}

\item{data}{the dataframe used in the regression that created the \code{vws} object}
}
\value{
the imput object is returned silently
}
\description{
\code{getxreg} gets the xreg that was used in a vws regression
}
\details{
Not much to add
}

