% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{Gumbel}
\alias{Gumbel}
\alias{r_gumbel}
\alias{d_gumbel}
\alias{p_gumbel}
\alias{q_gumbel}
\title{Gumbel Distribution}
\usage{
r_gumbel(n, mu = 0, sigma = 1)

d_gumbel(x, mu = 0, sigma = 1, log = FALSE)

p_gumbel(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

q_gumbel(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{Number of desired draws}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{x}{Vector of quantiles}

\item{log}{Logical; if \code{TRUE}, probabilities p are given as \eqn{log(p)}}

\item{q}{Vector of quantiles}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are
\eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}

\item{log.p}{Logical; if \code{TRUE}, probabilities p are given as \eqn{log(p)}}

\item{p}{Vector of probabilities}
}
\value{
A vector of draws
}
\description{
Functions for the Gumbel distribution
}
