% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlag.R
\name{mlag}
\alias{mlag}
\title{Creates a matrix of multiple lags (0:-p)}
\usage{
mlag(variable, lags)
}
\arguments{
\item{variable}{any variable}

\item{lags}{specified as from:to}
}
\value{
the imput object is returned silently
}
\description{
\code{mlag} returns VIF-values of the independent variables in the lag/diff form 
used in regression
}
\details{
It's really quite simple
}

