% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{MultivariateNormal}
\alias{MultivariateNormal}
\alias{r_mvnorm}
\alias{r_mvnorm_chol}
\alias{r_mvnorm_prec}
\alias{d_mvnorm}
\alias{d_mvnorm_prec}
\title{Multivariate Normal distribution}
\usage{
r_mvnorm(n, mu, Sigma)

r_mvnorm_chol(n, mu, Sigma_chol)

r_mvnorm_prec(n, mu, Omega)

d_mvnorm(x, mu, Sigma, log = FALSE)

d_mvnorm_prec(x, mu, Omega, log = FALSE)
}
\arguments{
\item{n}{Number of observations.}

\item{mu}{mean vector parameter.}

\item{Sigma}{Covariance matrix parameter.}

\item{Sigma_chol}{Cholesky factor of covariance matrix parameter.}

\item{Omega}{Precision matrix parameter.}

\item{x}{An \eqn{n \times k} matrix with n arguments.}

\item{log}{If \code{TRUE}, return densities and probabilities on the log-scale.}
}
\value{
\code{dmvnorm} gives the density using specified covariance matrix,
\code{dmvnorm_prec} gives the density using specified precision matrix,
\code{rmvnorm} generates random deviates using specified covariance matrix,
\code{rmvnorm_prec} generates random deviates using specified precision matrix,
\code{r_singular_mvnorm} generates random deviates using specified singular
covariance matrix (and mean zero).
}
\description{
k-dimensional Multivariate Normal distribution.
}
\details{
A use case for the function \code{r_singular_mvnorm} would be to start with
a singular precision matrix \code{Q}, and use something like
\code{Sigma = ginverse(Q)} as the covariance matrix.
}
