% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-helper.R
\name{normal_univariate_helper}
\alias{normal_univariate_helper}
\title{Normal Distribution Helper}
\usage{
normal_univariate_helper(mean, sd)
}
\arguments{
\item{mean}{Mean parameter \eqn{\mu}.}

\item{sd}{Standard deviation parameter \eqn{\sigma}.}
}
\description{
A distribution helper based on \eqn{\text{N}(\mu, \sigma^2)}.
}
\examples{
helper = normal_univariate_helper(mean = 0, sd = 1)
helper$d(0)
helper$p(1.96)
helper$q(0.025)

}
