% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimize_hybrid}
\alias{optimize_hybrid}
\title{Optimize Hybrid}
\usage{
optimize_hybrid(f, init, lower, upper, maximize, maxiter = 100000L)
}
\arguments{
\item{f}{TBD}

\item{init}{TBD}

\item{lower}{TBD}

\item{upper}{TBD}

\item{maximize}{TBD}

\item{maxiter}{TBD}
}
\description{
Optimize Hybrid
}
