% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vws}
\alias{plot.vws}
\title{Plots vws estimate results}
\usage{
\method{plot}{vws}(object, which = c(1, 2, 3), fromdate = NULL,
  todate = NULL, interactive = FALSE)
}
\arguments{
\item{object}{an object of class \code{vws}}

\item{which}{as in plots to show}

\item{fromdate}{starting date to plot from, has to be readable as date}

\item{todate}{ending date in plot}

\item{interactive}{if true lets you click on data points in plot 1 (actual)}
}
\value{
the imput object is returned silently
}
\description{
\code{plot.vws} plots fitted vs actual and residuals for regression with 
ARMA errors
}
\details{
By default 2 sets of plots are shown. First actual v.s. fitted and residuals.
Second two graphs showing the auto correaltion function (ACF) and 
the partial autocorrelation fuction.
}
\examples{
eu<-data.frame(EuStockMarkets)
eu$date<-seq.Date(from = as.Date('2014-01-30'), by = 'day', length.out = nrow(eu))
reg<-vws(diff(DAX, 1)~diff(CAC,1)+diff(SMI,1)+lag(diff(SMI,1),-5), ar=1, ma = 0, 
   data = subset(eu, subset = date>'2019-01-01'), datecolumn = "date")
plot(reg, which = 1, interactive=TRUE)
plot(reg, fromdate = "2019-02-15", todate = "2019-03-01")
}

