% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-helper.R
\name{poisson_univariate_helper}
\alias{poisson_univariate_helper}
\title{Poisson Distribution Helper}
\usage{
poisson_univariate_helper(lambda)
}
\arguments{
\item{lambda}{Rate parameter \eqn{\mu}.}
}
\description{
A distribution helper based on \eqn{\text{Poisson}(\lambda)}.
}
\examples{
helper = poisson_univariate_helper(lambda = 10)
helper$d(5)
helper$p(10)
helper$q(0.025)

}
