% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{polar}
\alias{polar}
\alias{inv_polar}
\title{Polar transformation}
\usage{
polar(z)

inv_polar(u)
}
\arguments{
\item{z}{A point in the unit sphere
\eqn{\mathbb{S}^d = \{ \bm{z} \in \mathbb{R}^d: \bm{z}^\top \bm{z} = 1 \}}.}

\item{u}{A point in the unit cube \eqn{[0,1]^{d-1}}.}
}
\description{
The polar transformation and its inverse, as expressed in Section 2.7 of
Anderson (2003), restricted to the unit sphere.
}
\details{
The \code{inv_polar} transformation \code{z = inv_polar(u)}, from
\eqn{[0,1]^{d-1}} to \eqn{\mathbb{S}^d}, can be described as a
composition of two transformations. The result
\deqn{
z_j =
\begin{cases}
\sin \phi_j \prod_{\ell=1}^{j-1} \cos \phi_\ell, & j = 1, \ldots, d-1, \\
\prod_{\ell=1}^{d-1} \cos \phi_\ell & j = d,
\end{cases}
}
is computed from angles
\deqn{
\phi_j =
\begin{cases}
\pi u_j - \pi / 2 , & j = 1, \ldots, d-2, \\
2 \pi u_j - \pi, & j = d-1.
\end{cases}
}

The \code{polar} transformation \code{u = polar(z)}, from \eqn{\mathbb{S}^d}
to \eqn{[0,1]^{d-1}}, is the inverse of \code{inv_polar}:
\deqn{
u_j =
\begin{cases}
(\phi_j + \pi / 2) / \pi, & j = 1, \ldots, d-2, \\
(\phi_j + \pi) / (2\pi), & j = d-1,
\end{cases}
}
and
\deqn{
\phi_j =
\begin{cases}
\arcsin \left\{ z_j \Big/ \sqrt{1 - z_1^2 - \cdots - z_{j-1}^2} \right\}, & j = 1, \ldots, d-1, \\
\arctan\{ z_{d-1} / z_d \}, & j = d.
\end{cases}
}
}
\examples{
x = 1:5 / sum(1:5)
z = inv_polar(x)
polar(z)

}
\references{
T.W. Anderson (2003). An Introduction to Multivariate
Statistical Analysis, Wiley-Interscience, 3rd Edition.
}
