% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.vws}
\alias{predict.vws}
\title{Predict using a fitted vws model}
\usage{
\method{predict}{vws}(object, newxreg = NULL, se.fit = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{vws}}

\item{se.fit}{return standard errors or not}

\item{n.ahead}{number of periods ahead to predict}

\item{data}{the original data}
}
\value{
the imput object is returned silently
}
\description{
\code{predict.vws} Returns predicted values and standard errors for new data 
ARMA errors
}
\details{
Supplied a vws model and a new dataset it predicts n periods ahead. Not that n must
equal the number of rows in the new dataset.
}
\examples{
eu<-data.frame(EuStockMarkets)
eu$date<-seq.Date(from = as.Date('2014-01-30'), by = 'day', length.out = nrow(eu))
reg<-vws(diff(DAX, 1)~diff(CAC,1)+diff(SMI,1)+lag(diff(SMI,1),-5), ar=1, ma = 0, 
   data = subset(eu, subset = date>'2019-01-01'), datecolumn = "date")
predict.vws(reg, n.ahead = 13, dataset = eu, newxreg = getxreg(reg, subset(eu, date>="2019-02-20")))
}

