% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predplot.R
\name{predplot}
\alias{predplot}
\title{Plots the prediction of a vws model with confidence intervals}
\usage{
predplot(object = NULL, prediction = NULL)
}
\arguments{
\item{object}{an object of class \code{vws}}

\item{prediction}{as returned from vws.predict}
}
\value{
the imput object is returned silently
}
\description{
\code{predplot} plots fitted vs actual and residuals for regression with 
ARMA errors
}
\details{
Similar to plot(forecast())) in forecast library
}

