% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectConstRegion.R
\name{RectConstRegion}
\alias{RectConstRegion}
\title{Univariate Region with Constant Majorizer}
\description{
Univariate Region with Constant Majorizer

Univariate Region with Constant Majorizer
}
\details{
An R6 class which represents a region based on univariate intervals with a
constant majorizer for the weight function. This version is for continuous
supports.
}
\examples{
# Define base distribution and weight function
g = normal_univariate_helper(mean = 0, sd = 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }

reg = RectConstRegion$new(-Inf, 10, w, g)
print(reg)

out = reg$bifurcate(0)
print(out[[1]])
print(out[[2]])

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{w}}{Weight function for the target distribution. Its expected interface
is \code{w(x, log = TRUE)} so that results are returned on the log-scale by
default.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RectConstRegion-new}{\code{RectConstRegion$new()}}
\item \href{#method-RectConstRegion-d_base}{\code{RectConstRegion$d_base()}}
\item \href{#method-RectConstRegion-r}{\code{RectConstRegion$r()}}
\item \href{#method-RectConstRegion-d}{\code{RectConstRegion$d()}}
\item \href{#method-RectConstRegion-s}{\code{RectConstRegion$s()}}
\item \href{#method-RectConstRegion-w_major}{\code{RectConstRegion$w_major()}}
\item \href{#method-RectConstRegion-midpoint}{\code{RectConstRegion$midpoint()}}
\item \href{#method-RectConstRegion-bifurcate}{\code{RectConstRegion$bifurcate()}}
\item \href{#method-RectConstRegion-bifurcate_at}{\code{RectConstRegion$bifurcate_at()}}
\item \href{#method-RectConstRegion-is_bifurcatable}{\code{RectConstRegion$is_bifurcatable()}}
\item \href{#method-RectConstRegion-xi_upper}{\code{RectConstRegion$xi_upper()}}
\item \href{#method-RectConstRegion-xi_lower}{\code{RectConstRegion$xi_lower()}}
\item \href{#method-RectConstRegion-upper}{\code{RectConstRegion$upper()}}
\item \href{#method-RectConstRegion-lower}{\code{RectConstRegion$lower()}}
\item \href{#method-RectConstRegion-description}{\code{RectConstRegion$description()}}
\item \href{#method-RectConstRegion-print}{\code{RectConstRegion$print()}}
\item \href{#method-RectConstRegion-optimize}{\code{RectConstRegion$optimize()}}
\item \href{#method-RectConstRegion-clone}{\code{RectConstRegion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-new"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$new(a, b, w, g)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{Lower limits of intervals.}

\item{\code{b}}{Upper limits of intervals.}

\item{\code{w}}{Weight function for the target distribution.}

\item{\code{g}}{An list of objects created by \code{univariate_helper}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-d_base"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-d_base}{}}}
\subsection{Method \code{d_base()}}{
Density function \eqn{g} for the base distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$d_base(x, log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Density argument.}

\item{\code{log}}{logical; if \code{TRUE}, return result on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-r"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-r}{}}}
\subsection{Method \code{r()}}{
Generate a draw from \eqn{g_j} specific to this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$r(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of draws to generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of draws, with one draw per list element.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-d"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-d}{}}}
\subsection{Method \code{d()}}{
Density of \eqn{g_j} specific to this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$d(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Density argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-s"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-s}{}}}
\subsection{Method \code{s()}}{
Test if given \code{x} is in the support for the \eqn{g_j} specific to this
region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$s(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Density argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-w_major"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-w_major}{}}}
\subsection{Method \code{w_major()}}{
Majorized weight function \eqn{\overline{w}_j} for this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$w_major(x, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Argument to weight function.}

\item{\code{log}}{logical; if \code{TRUE}, return result on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-midpoint"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-midpoint}{}}}
\subsection{Method \code{midpoint()}}{
Compute a midpoint for the region. This is defined to be the standard
midpoint for regions with finite limits; otherwise we select a point between
the bounds.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$midpoint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-bifurcate"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-bifurcate}{}}}
\subsection{Method \code{bifurcate()}}{
Bifurcate this region into two regions at the midpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$bifurcate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-bifurcate_at"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-bifurcate_at}{}}}
\subsection{Method \code{bifurcate_at()}}{
Bifurcate this region into two regions by splitting dimension \code{l} at
point \code{x}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$bifurcate_at(l, x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{An integer \eqn{l \in \{ 1, \ldots, k\} }; which dimension to split.}

\item{\code{x}}{A scalar.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-is_bifurcatable"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-is_bifurcatable}{}}}
\subsection{Method \code{is_bifurcatable()}}{
Return a logical value indicating whether this region is bifurcatable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$is_bifurcatable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-xi_upper"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-xi_upper}{}}}
\subsection{Method \code{xi_upper()}}{
The quantity \eqn{\overline{\xi}_j} for this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$xi_upper(log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log}}{logical; if \code{TRUE}, return result on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-xi_lower"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-xi_lower}{}}}
\subsection{Method \code{xi_lower()}}{
The quantity \eqn{\underline{\xi}_j} for this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$xi_lower(log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log}}{logical; if \code{TRUE}, return result on the log-scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-upper"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-upper}{}}}
\subsection{Method \code{upper()}}{
The upper limits \eqn{\alpha_j} for this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$upper()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-lower"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-lower}{}}}
\subsection{Method \code{lower()}}{
The lower limits \eqn{\alpha_{j-1}} for this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$lower()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-description"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-description}{}}}
\subsection{Method \code{description()}}{
A string that describes the region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$description()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-print"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-optimize}{}}}
\subsection{Method \code{optimize()}}{
Maximize or minimize the function \eqn{w(x)} over this region. Optimization
is carried out with \code{optim} using arguments
\code{method = "Nelder-Mead"} and
\code{control = list(maxit = 100000, warn.1d.NelderMead = FALSE)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$optimize(maximize = TRUE, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maximize}}{logical; if \code{TRUE} do maximization. Otherwise do
minimization.}

\item{\code{log}}{logical; if \code{TRUE} return optimized value of \eqn{\log w(x)}.
Otherwise return optimized value of \eqn{w(x)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RectConstRegion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RectConstRegion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RectConstRegion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
