% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{Region}
\alias{Region}
\title{Region}
\description{
A region contains all of the problem-specific logic for the vws sampler.
This is an "abstract" region that shows which functions need to be
implemented.
}
