% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection.R
\name{rejection}
\alias{rejection}
\title{Vertical Weighted Strips Rejection Sampler}
\usage{
rejection(h, n = 1, control = rejection_control())
}
\arguments{
\item{h}{An \code{fmm_proposal} object}

\item{n}{Number of desired draws}

\item{control}{A control object from \code{rejection_control}}
}
\value{
A list whose structure depends on
\code{extra_outputs}. If \code{extra_outputs = FALSE}, the list is of length
\code{n} where each element represents one draw. If
\code{extra_outputs = TRUE}, the list contains the following named elements:

\item{draws}{is list is of length \code{n} where each element represents one
draw.}
\item{rejects}{is a vector of counts. The \eqn{i}th element is the number of
rejections before the \eqn{i}th successful draw.}
\item{h}{returns the \code{h} that was input to the sampler.}

Note that \code{h} was originally intended to show adaptation during
sampling, but that is currently not done.
}
\description{
Accept-reject algorithm using our proposal for weighted distributions.
}
\examples{
# Define base distribution and weight function
g = normal_univariate_helper(0, 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }

# Set up support
support = univariate_const_region(-Inf, 10, w, g)
regions = support$bifurcate()

# Create a finite mixture proposal
h = fmm_proposal(regions)
h$rejection_bound()
h$rejection_bound(byregion = TRUE)

out = rejection(h, n = 1000)
print(out |> unlist())

out = rejection(h, n = 1000, rejection_control(extra_outputs = TRUE))
print(out$draws |> unlist())
print(out$rejects)

}
