% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection-bound.R
\name{rejection_bound}
\alias{rejection_bound}
\title{Bound for probability of rejection}
\usage{
rejection_bound(log_xi_upper, log_xi_lower, log = FALSE)
}
\arguments{
\item{log_xi_upper}{The vector \eqn{\log \overline{\xi}_1, \ldots, \log \overline{\xi}_N}}

\item{log_xi_lower}{The vector \eqn{\log \underline{\xi}_1, \ldots, \log \underline{\xi}_N}}

\item{log}{If \code{TRUE}, return result on the log-scale}
}
\description{
Compute the expression
\deqn{
1 - \sum_{j=1}^N \underline{\xi}_j \Bigg/ \sum_{j=1}^N \overline{\xi}_j
}
in a numerically stable way.
}
