% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection-control.R
\name{rejection_control}
\alias{rejection_control}
\title{Rejection Control}
\usage{
rejection_control(
  N = 30,
  tol = 0.1,
  max_rejects = .Machine$integer.max,
  report = .Machine$integer.max,
  ratio_ub = exp(1e-05),
  action = c("stop", "warning", "message", "none")
)
}
\arguments{
\item{N}{Maximum number of regions to create in the proposal.}

\item{tol}{Desired bound for probability of rejection in the proposal.}

\item{max_rejects}{Maximum number of rejections to tolerate before bailing out.}

\item{report}{Report progress each time this many candidates are proposed.}

\item{ratio_ub}{TBD}

\item{action}{What should happen if sampler halts with \code{max_rejects}
rejections: one of \code{"stop"},  \code{"warning"}, or \code{"message"}.}
}
\value{
A control object to be passed to the \code{rejection} function.
}
\description{
Control object for rejection sampler.
}
