% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{RUnivariateHelper}
\alias{RUnivariateHelper}
\title{RUnivariateHelper}
\description{
Export the "RUnivariateHelper" C++ class by explicitly requesting
RUnivariateHelper be exported via roxygen2's export tag.
}
