% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.vws}
\alias{summary.vws}
\title{Summarizes vws estimate results}
\usage{
\method{summary}{vws}(object)
}
\arguments{
\item{object}{an object of class \code{vws}}
}
\value{
the imput object is returned silently
}
\description{
\code{summary.vws} summarizes the results of a regression with ARMA errors
}
\details{
Shows the most important information about the regression
}
\examples{
#'eu<-data.frame(EuStockMarkets)
eu$date<-seq.Date(from = as.Date('2014-01-30'), by = 'day', length.out = nrow(eu))
reg<-vws(diff(DAX, 1)~diff(CAC,1)+diff(SMI,1)+lag(diff(SMI,1),-5), ar=1, ma = 0, 
   data = subset(eu, subset = date>'2019-01-01'), datecolumn = "date")
summary(reg)
}

