% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TExp}
\alias{TExp}
\alias{n_texp}
\alias{integrate_texp}
\alias{d_texp}
\alias{p_texp}
\alias{q_texp}
\alias{r_texp}
\alias{mgf_texp}
\title{Truncated Exponential}
\usage{
n_texp(kappa, lo, hi, log = FALSE)

integrate_texp(a, b, kappa, lo, hi, log = FALSE)

d_texp(x, kappa, lo, hi, log = FALSE)

p_texp(q, kappa, lo, hi, lower = TRUE, log = FALSE)

q_texp(p, kappa, lo, hi, lower = TRUE, log = FALSE)

r_texp(n, kappa, lo, hi)

mgf_texp(s, kappa, lo, hi, log = FALSE)
}
\description{
Truncated Exponential
}
