% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-helper.R
\name{uniform_helper}
\alias{uniform_helper}
\title{Uniform Distribution Helper}
\usage{
uniform_helper(a, b)
}
\arguments{
\item{a}{lower limit of support.}

\item{b}{upper limit of support.}
}
\description{
A distribution helper based on \eqn{\text{Uniform}(a, b)}.
}
\examples{
helper = uniform_helper(a = 0, b = 1)
helper$d(1/2)
helper$p(1/4)
helper$q(0.025)

}
