% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-const-region.R
\name{UnivariateConstRegion}
\alias{UnivariateConstRegion}
\alias{univariate_const_region}
\title{Univariate Region with Constant Majorizer}
\usage{
univariate_const_region(a, b, w, g)
}
\arguments{
\item{a}{Lower knot of region.}

\item{b}{Upper knot of region.}

\item{w}{Weight function.}

\item{g}{Object that encapsulates base distribution.}

\item{x}{Object of type \code{univariate_const_region}.}

\item{...}{additional arguments.}
}
\description{
A region based on univariate intervals with a constant majorizer for the
weight function.
}
\examples{
#  Define base distribution
g = univariate_helper(
    r = function(n) rnorm(n, 0, 5),
    d = function(x, log = FALSE) dnorm(x, 0, 5, log),
    p = function(q, lower.tail = TRUE, log.p = FALSE) {
        pnorm(q, 0, 5, lower.tail, log.p)
    },
    q = function(p, lower.tail = TRUE, log.p = FALSE) {
        qnorm(p, 0, 5, lower.tail, log.p)
    },
    in_support = function(x) { TRUE }
)

# Define weight function
w = function(x, log = FALSE) {
    dlnorm(10 - x, 5, 2, log)
}

reg = univariate_const_region(-Inf, 10, w, g)
print(reg)

out = reg$bifurcate(0)
print(out[[1]])
print(out[[2]])

}
