% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{get VIF-values of independent variables}
\usage{
vif(object)
}
\arguments{
\item{object}{an object of class \code{vws}}

\item{data}{the dataframe used in the regression that created the \code{vws} object}
}
\value{
the imput object is returned silently
}
\description{
\code{vif} returns VIF-values of the independent variables in the lag/diff form 
used in regression
}
\details{
It's really quite simple
}

