% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vws-package}
\alias{vws}
\alias{vws-package}
\title{vws}
\description{
Package documentation
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewraim/vws}
}

}
\author{
\strong{Maintainer}: Andrew M. Raim \email{andrew.raim@gmail.com}

}
