% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\title{Adapt Using Midpoint}
\usage{
adapt(h, N, tol = -Inf, report = N + 1)
}
\arguments{
\item{h}{An object of class \code{FMMProposal}.}

\item{N}{Number of additional mixture components after adaptation.}

\item{tol}{A scalar; adaptation halts if the log of the rejection bound is
smaller than this value. The default value \code{-Inf} prevents early halting.}

\item{report}{Report progress each time this many candidates are proposed.}
}
\description{
Adapt an FMM proposal using a midpoint rule-of-thumb.
}
\examples{
g = normal_univariate_helper(0, 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }
support = UnivariateConstRegion$new(-Inf, 10, w, g)
regions = list(support)

# Adapt the proposal
h_init = FMMProposal$new(regions)
adapt_out = adapt(h_init, N = 100)
h = adapt_out$h

# Create a finite mixture proposal
h$rejection_bound()
h$rejection_bound(byregion = TRUE)

out = rejection(h, n = 200)
print(out |> unlist())

out = rejection(h, n = 200, rejection_control(extra_outputs = TRUE))
print(out$draws |> unlist())
print(out$rejects)

}
