% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal-spaced-knots.R
\name{equal_spaced_knots}
\alias{equal_spaced_knots}
\title{Equally Spaced Univariate Knots}
\usage{
equal_spaced_knots(N, lo, hi)
}
\arguments{
\item{N}{Number of desired intervals.}

\item{lo}{Left endpoint; must be finite.}

\item{hi}{Right endpoint; must be finite.}
}
\value{
A vector of \eqn{N-1} knots.
}
\description{
Prepare knots which define \eqn{N} equally-spaced intervals between finite
endpoints \code{lo} and \code{hi}.
}
\examples{
equal_spaced_knots(2, lo = 0, hi = 1)
equal_spaced_knots(3, lo = 0, hi = 1)

}
