% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMMProposal.R
\name{FMMProposal}
\alias{FMMProposal}
\alias{fmm_proposal}
\title{FMM Proposal}
\usage{
fmm_proposal(regions)
}
\arguments{
\item{regions}{A list of regions that form a partition of the support.}
}
\description{
FMM Proposal
}
\details{
The \code{regions} list represents \eqn{\mathcal{D}_1, \ldots, \mathcal{D}_N}.
They should be objects based on a subclass of \code{Region}, which uses
R's reference class construct.
}
\section{Fields}{

\describe{
\item{\code{regions}}{A list of \code{N} regions that form a partition of the support.}

\item{\code{log_xi_upper}}{The numeric vector
\eqn{\overline{\xi}_1, \ldots, \overline{\xi}_N}.}

\item{\code{log_xi_lower}}{The numeric vector
\eqn{\underline{\xi}_1, \ldots, \underline{\xi}_N}.}

\item{\code{bifurcatable}}{A vector of logical values which indicates whether the
corresponding regions may be bifurcated further.}
}}

\examples{
# Define base distribution and weight function
g = normal_univariate_helper(mean = 0, sd = 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }

# Set up support
support = univariate_const_region(-Inf, 10, w, g)
regions = support$bifurcate()

# Create a finite mixture with one component
fmm = fmm_proposal(regions)
print(fmm)

}
