% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm-proposal.R
\docType{class}
\name{FMMProposal}
\alias{FMMProposal}
\alias{fmm_proposal}
\title{FMM Proposal}
\usage{
fmm_proposal(regions)
}
\arguments{
\item{regions}{A list of regions that form a partition of the support}

\item{x}{An FMM Proposal to operate on}

\item{object}{An FMM Proposal to operate on}

\item{n}{Number of regions to}

\item{...}{Additional arguments}
}
\description{
Use reference classes here to allow proposal to be mutable.
}

\examples{
#  Define base distribution
g = univariate_helper(
    r = function(n) rnorm(n, 0, 5),
    d = function(x, log = FALSE) dnorm(x, 0, 5, log),
    p = function(q, lower.tail = TRUE, log.p = FALSE) {
        pnorm(q, 0, 5, lower.tail, log.p)
    },
    q = function(p, lower.tail = TRUE, log.p = FALSE) {
        qnorm(p, 0, 5, lower.tail, log.p)
    },
    in_support = function(x) { TRUE }
)

# Define weight function
w = function(x, log = FALSE) {
    dlnorm(10 - x, 5, 2, log)
}

# Set up support
support = univariate_const_region(-Inf, 10, w, g)
regions = support$bifurcate()

# Create a finite mixture with one component
fmm = fmm_proposal(regions)
print(fmm)

}
