% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntUnivariateConstRegion.R
\name{IntUnivariateConstRegion}
\alias{IntUnivariateConstRegion}
\title{Integer Univariate Region with Constant Majorizer}
\description{
Integer Univariate Region with Constant Majorizer

Integer Univariate Region with Constant Majorizer
}
\details{
A region based on univariate intervals with a constant majorizer for the
weight function. This version is for integer supports.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{w}}{Weight function for the target distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnivariateConstRegion-new}{\code{IntUnivariateConstRegion$new()}}
\item \href{#method-UnivariateConstRegion-r}{\code{IntUnivariateConstRegion$r()}}
\item \href{#method-UnivariateConstRegion-bifurcate}{\code{IntUnivariateConstRegion$bifurcate()}}
\item \href{#method-UnivariateConstRegion-is_bifurcatable}{\code{IntUnivariateConstRegion$is_bifurcatable()}}
\item \href{#method-UnivariateConstRegion-print}{\code{IntUnivariateConstRegion$print()}}
\item \href{#method-UnivariateConstRegion-clone}{\code{IntUnivariateConstRegion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$new(a, b, w, g, log_w_max, log_w_min)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{Lower limit of interval.}

\item{\code{b}}{Upper limit of interval.}

\item{\code{w}}{Weight function for the target distribution.}

\item{\code{g}}{An object created by \code{univariate_helper}.}

\item{\code{log_w_max}}{The value \eqn{\max_{x \in (a,b]}\log w(x)}.}

\item{\code{log_w_min}}{The value \eqn{\min_{x \in (a,b]}\log w(x)}.}

\item{\code{log_prob}}{The value \eqn{\log \text{P}(a < T \leq b)} for \eqn{T \sim g}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-r"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-r}{}}}
\subsection{Method \code{r()}}{
Generate a draw from \eqn{g_j} specific to this region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$r(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of draws to generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of draws, with one draw per list element.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-bifurcate"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-bifurcate}{}}}
\subsection{Method \code{bifurcate()}}{
Bifurcate this region into two regions. Use \code{x} as the bifurcation
point if it is not \code{NULL}. Otherwise, select a point for bifurcation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$bifurcate(x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An optional bifurcation point.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-is_bifurcatable"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-is_bifurcatable}{}}}
\subsection{Method \code{is_bifurcatable()}}{
Return a logical value indicating whether this region is bifurcatable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$is_bifurcatable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-print"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnivariateConstRegion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnivariateConstRegion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntUnivariateConstRegion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
