% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int-univariate-const-region.R
\name{IntUnivariateConstRegion}
\alias{IntUnivariateConstRegion}
\alias{int_univariate_const_region}
\title{Integer Univariate Region with Constant Majorizer}
\usage{
int_univariate_const_region(a, b, w, g)
}
\arguments{
\item{a}{Lower knot of region.}

\item{b}{Upper knot of region.}

\item{w}{Weight function.}

\item{g}{Object that encapsulates base distribution.}

\item{s}{Object of type \code{univariate_const_region}.}

\item{x}{Object of type \code{univariate_const_region}.}

\item{...}{additional arguments.}
}
\description{
A region based on univariate intervals with a constant majorizer for the
weight function. This version is for integer supports.
}
\examples{
#  Define base distribution
g = univariate_helper(
    r = function(n) rpois(n, 5),
    d = function(x, log = FALSE) dpois(x, 5, log),
    p = function(q, lower.tail = TRUE, log.p = FALSE) {
        ppois(q, 5, lower.tail, log.p)
    },
    q = function(p, lower.tail = TRUE, log.p = FALSE) {
        qpois(p, 5, lower.tail, log.p)
    },
    in_support = function(x) { is.integer(x) & x >= 0 }
)

# Define weight function
w = function(x, log = FALSE) {
    dlnorm(10 - x, 5, 2, log)
}

reg = int_univariate_const_region(-Inf, 10, w, g)
print(reg)

out = reg$bifurcate(0)
print(out[[1]])
print(out[[2]])

}
