% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InverseGamma}
\alias{InverseGamma}
\alias{r_invgamma}
\alias{d_invgamma}
\alias{p_invgamma}
\alias{q_invgamma}
\title{Inverse Gamma distribution}
\usage{
r_invgamma(n, a, b)

d_invgamma(x, a, b, log = FALSE)

p_invgamma(q, a, b, lower = TRUE, log = FALSE)

q_invgamma(p, a, b, lower = TRUE, log = FALSE)
}
\arguments{
\item{n}{Number of draws.}

\item{a}{Shape parameter.}

\item{b}{Rate parameter.}

\item{x}{Vector of quantiles.}

\item{log}{If \code{TRUE}, return densities and probabilities on the log-scale.}

\item{q}{Vector of quantiles.}

\item{lower}{logical; if TRUE (default), probabilities are
\eqn{P(X \leq x)}; otherwise, \eqn{P(X > x)}.}

\item{p}{Vector of probabilities.}
}
\value{
\code{d_invgamma} computes the density, \code{r_invgamma} generates random deviates,
\code{p_invgamma} computes the CDF, and \code{q_invgamma} computes quantiles.
}
\description{
Inverse Gamma distribution
}
\details{
Note that \verb{Rcpp::*gamma} and \verb{R::*gamma} functions are both parameterized
by a scale parameter, which is the inverse of the rate.
}
