% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{InverseGamma}
\alias{InverseGamma}
\alias{r_invgamma}
\alias{d_invgamma}
\alias{p_invgamma}
\alias{q_invgamma}
\title{Inverse Gamma distribution}
\usage{
r_invgamma(n, a, b)

d_invgamma(x, a, b, log = FALSE)

p_invgamma(q, a, b, lower.tail = TRUE, log.p = FALSE)

q_invgamma(p, a, b, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{Number of observations.}

\item{a}{Shape parameter.}

\item{b}{Rate parameter.}

\item{x}{Vector of quantiles.}

\item{log}{If \code{TRUE}, return densities and probabilities on the log-scale.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P(X \leq x)}; otherwise, \eqn{P(X > x)}.}

\item{log.p}{If \code{TRUE}, input and output probabilities on the log-scale.}

\item{p}{Vector of probabilities.}
}
\value{
\code{dinvgamma} gives the density, \code{rinvgamma} generates random
deviates.
}
\description{
Inverse Gamma distribution
}
