% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sum_exp}
\alias{log_sum_exp}
\alias{log_sum_exp_mat}
\alias{log_add2_exp}
\alias{log_sub2_exp}
\title{Log-Sum-Exp}
\usage{
log_sum_exp(x)

log_sum_exp_mat(x)

log_add2_exp(x, y)

log_sub2_exp(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\description{
Compute \code{log(sum(exp(x)))} in a more stable way.
}
\details{
Computed using the method described by user Ben in StackExchange
thread \url{https://stats.stackexchange.com/questions/381936/vectorised-computation-of-logsumexp}.

The function \code{log_sub2_exp} expects that each element of \code{x} is
larger than or equal to its corresponding element in \code{y}. Otherwise,
\code{NaN} will be returned with a warning.
}
\examples{
pi = 1:6 / sum(1:6)
x = log(2*pi)
log(sum(exp(x)))
log_sum_exp(x)

# Result should be 0
x = c(-Inf -Inf, 0)
log_sum_exp(x)

# Result should be -Inf
x = c(-Inf -Inf, -Inf)
log_sum_exp(x)

# Result should be Inf
x = c(-Inf -Inf, Inf)
log_sum_exp(x)

# Result should be 5 on the original scale
out = log_add2_exp(log(3), log(2))
exp(out)

# Result should be 7 on the original scale
out = log_sub2_exp(log(12), log(5))
exp(out)

}
