% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\title{Region}
\description{
A region contains all of the problem-specific logic for the vws sampler.
This is an "abstract" region that shows which functions need to be
implemented.
}
\details{
The following methods should be implemented:
\describe{
\item{\code{d_base}}{Density function \eqn{g} for the base distribution.}
\item{\code{w}}{Weight function \eqn{w}.}
\item{\code{r}}{Generate a draw from \eqn{g_j} specific to this region.}
\item{\code{d}}{Density of \eqn{g_j} specific to this region.}
\item{\code{in_support}}{Test if given \code{x} is in the support for the
\eqn{g_j} specific to this region.}
\item{\code{w_major}}{Majorized weight function \eqn{\overline{w}_j} for
this region.}
\item{\code{bifurcate}}{	Bifurcate this region into two regions. Use
\\code{x} as the bifurcation point if it is not \code{NULL}. Otherwise,
select a point for bifurcation.}
\item{\code{xi_upper}}{The quantity \eqn{\overline{\xi}_j} for this
region.}
\item{\code{xi_lower}}{The quantity \eqn{\underline{\xi}_j} for this
region.}
}
}
