% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection.R
\name{rejection}
\alias{rejection}
\alias{rejection_numeric}
\alias{rejection_int}
\title{Rejection Sampling with VWS Proposal}
\usage{
rejection_numeric(n, lo, hi, w, helper, control = rejection_control())

rejection_int(n, lo, hi, w, helper, control = rejection_control())
}
\arguments{
\item{n}{Number of desired draws.}

\item{lo}{Lower bound of support.}

\item{hi}{Upper bound of support.}

\item{w}{The weight function \eqn{w}.}

\item{helper}{A helper object constructed from \link{univariate_helper} that
characterizes the base distribution \eqn{g}.}

\item{control}{An object constructed using \link{rejection_control}.}
}
\value{
A list with the following elements.
\itemize{
\item \code{draws}: Vector with accepted draws.
\item \code{rejects}: Vector with counts of rejections needed to obtain each draw.
\item \code{lbdd}: Value of bound for the log-probability of rejection over
refinement of the proposal. The maximum length of this vector is \eqn{N},
but it may be shorter if tolerance \code{tol} was achieved before partitioning
into \eqn{N} regions.
}

The length of \code{draws} and \code{rejects} will be \eqn{n} unless the maximum
number of rejections was exceeded.
}
\description{
Rejection sampling for univariate targets using a constant majorizer.
Constants are obtained using numerical optimization. The proposal is created
internally, used, and then discarded. See the C++ framework if access to the
proposal is needed or other variations of VWS are desired.
}
\details{
VWS requires decomposition of the target \eqn{f} into
\eqn{f(x) \propto w(x) g(x)}.

The functions \code{rejection_numeric} and \code{rejection_int} are for targets with
continuous and integer supports, respectively.
}
\examples{
library(tidyverse)
set.seed(1234)

z = 66
mu = 5
sigma = 0.25
lambda2 = 100
tol = 0.2

df = \(x, log = F) { dnorm(x, mean = z, sd = sqrt(lambda2), log) }
pf = \(q, lower.tail = T, log.p = F) {
  pnorm(q, mean = z, sd = sqrt(lambda2), lower.tail, log.p)
}
qf = \(p, lower.tail = T, log.p = F) {
  qnorm(p, mean = z, sd = sqrt(lambda2), lower.tail, log.p)
}
helper = univariate_helper(df, pf, qf)

w = function(y, log = TRUE) {
  dlnorm(y, mu, sigma, log = log)
}

ctrl = rejection_control(N = 50, tol = tol, max_rejects = 10000,
  action = "stop")
out = rejection_numeric(n = 10000, lo = 0, hi = Inf, w, helper,
  control = ctrl)
sum(out$rejects)

# Some plots of the results
data.frame(x = out$draws) \%>\%
  ggplot() +
  geom_histogram(aes(x = x), col = "black", bins = 20) +
  theme_light()

data.frame(draw = seq_along(out$lbdd), lbdd = out$lbdd) \%>\%
  ggplot() +
  geom_line(aes(x = draw, y = lbdd)) +
  geom_hline(yintercept = log(tol), lty = 2) +
  xlab("Draw") +
  ylab("Log of Bound for Rejection Probability") +
  theme_light()

data.frame(draw = seq_along(out$rejects), crej = cumsum(out$rejects)) \%>\%
  ggplot() +
  geom_line(aes(x = draw, y = crej)) +
  xlab("Draw") +
  ylab("Cumulative Rejection Count") +
  theme_light()

print("TBD: add an example with integer support")

}
