% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection-control.R
\name{rejection_control}
\alias{rejection_control}
\title{Rejection Control}
\usage{
rejection_control(
  max_rejects = 1000,
  report = 100,
  extra_outputs = FALSE,
  action_incomplete = c("stop", "warning", "message")
)
}
\arguments{
\item{max_rejects}{Maximum number of rejections to tolerate before bailing out.}

\item{report}{Report progress each time this many candidates are proposed.}

\item{extra_outputs}{If \code{TRUE}, return a list with extended output
in addition to the accepted draws. Otherwise only return accepted draws.}

\item{action_incomplete}{What should happen if sampler halts with
\code{max_rejects} rejections: one of \code{"stop"},  \code{"warning"}, or
\code{"message"}.}
}
\value{
A control object to be passed to the \code{rejection} function.
}
\description{
Control object for rejection sampler.
}
