% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-helper.R
\name{univariate_helper}
\alias{univariate_helper}
\title{Univariate Distribution Helper}
\usage{
univariate_helper(r, d, p, q, in_support)
}
\arguments{
\item{r}{Drawing function for base distribution.}

\item{d}{Density function for base distribution.}

\item{p}{CDF function for base distribution.}

\item{q}{Quantile function for base distribution.}

\item{in_support}{Support inclusion function for base distribution.}
}
\description{
TBD: document expected function interface.
}
