% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-helper.R
\name{univariate_helper}
\alias{univariate_helper}
\title{Univariate Distribution Helper}
\usage{
univariate_helper(d, p, q, in_support)
}
\arguments{
\item{d}{Density function for base distribution.}

\item{p}{Cumulative distribution function for base distribution.}

\item{q}{Quantile function for base distribution.}

\item{in_support}{Support inclusion function for base distribution.}
}
\description{
A container to support VWS with univariate distributions.
}
\examples{
univariate_helper(
    d = function(x, log = FALSE) dpois(x, log),
    p = function(q, lower.tail = TRUE, log.p = FALSE) {
        ppois(q, lower.tail, log.p)
    },
    q = function(p, lower.tail = TRUE, log.p = FALSE) {
        qpois(p, lower.tail, log.p)
    },
    in_support = function(x) { TRUE }
)

}
