% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateConstRegion.R
\name{UnivariateConstRegion}
\alias{UnivariateConstRegion}
\alias{univariate_const_region}
\title{Univariate Region with Constant Majorizer}
\usage{
univariate_const_region(a, b, w, g)
}
\arguments{
\item{a}{Lower knot of region.}

\item{b}{Upper knot of region.}

\item{w}{Weight function.}

\item{g}{Object that encapsulates base distribution.}
}
\description{
A region based on univariate intervals with a constant majorizer for the
weight function.
}
\examples{
# Define base distribution and weight function
g = normal_univariate_helper(mean = 0, sd = 5)
w = function(x, log = FALSE) { dlnorm(10 - x, meanlog = 5, sdlog = 2, log) }

reg = univariate_const_region(-Inf, 10, w, g)
print(reg)

out = reg$bifurcate(0)
print(out[[1]])
print(out[[2]])

}
